fix funtion type issue

+++ app/src/main/java/com/smarteq/services/AudioProcessingService.kt
@@ -186,7 +186,7 @@ class AudioProcessingService : Service() {
         }
         
         for (i in levels.indices) {
-            equalizer?.setBandLevel(i.toShort(), (levels[i] * 100).toShort())
+            equalizer?.setBandLevel(i.toInt().toShort(), (levels[i] * 100).toInt().toShort())
         }
     }
